finish chapter 10 ex

+++ test/integration/users_index_test.rb
@@ -5,6 +5,7 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
   def setup
     @admin     = users(:michael)
     @non_admin = users(:archer)
+    @inactive = users(:lana)
   end
 
   test "index as admin including pagination and delete links" do
@@ -14,10 +15,13 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
     assert_select 'div.pagination'
     first_page_of_users = User.paginate(page: 1)
     first_page_of_users.each do |user|
-      assert_select 'a[href=?]', user_path(user), text: user.name
-      unless user == @admin
-        assert_select 'a[href=?]', user_path(user), text: 'delete',
-                                                    method: :delete
+      unless !user.activated?
+        assert_select 'a[href=?]', user_path(user), text: user.name
+      
+        unless user == @admin
+          assert_select 'a[href=?]', user_path(user), text: 'delete',
+                                                      method: :delete
+        end
       end
     end
     assert_difference 'User.count', -1 do
